项目模块

+++ src/main/java/cn/com/partical/development/system/developmentservice/entity/ProjectMember.java
@@ -12,16 +12,24 @@ import lombok.EqualsAndHashCode;
 @EqualsAndHashCode(callSuper = false)
 public class ProjectMember extends BasicsEntity {
 
-  // 项目ID
+  /**
+   * 项目ID
+   */
   private long projectId;
 
-  // 用户ID
+  /**
+   * 用户ID
+   */
   private long userId;
 
-  // 项目成员类型 0：创建者，1：管理员，2：编辑员，3：阅读员
-  private byte projectMemberType;
+  /**
+   * 项目成员类型 0：创建者，1：管理员，2：编辑员，3：阅读员
+   */
+  private Integer projectMemberType;
 
-  // 项目成员状态 0：正常，1：冻结
+  /**
+   * 项目成员状态 0：正常，1：冻结
+   */
   private byte projectMemberState;
 
   // 标注 0: 未标注，1: 标注
